added global exception handles and also resolved the pom files for ve…

+++ Create-Task-Application/src/main/java/com/batch5/Create_Task_Application/UserModule/controller/UserRoleController.java
@@ -28,13 +28,13 @@ public class UserRoleController {
 
     //  GET ROLE BY ID
     @GetMapping("/{roleId}")
-    public UserRole getRoleById(@PathVariable Long roleId) {
+    public UserRole getRoleById(@PathVariable Integer roleId) {
         return roleService.getRoleById(roleId);
     }
 
     //  DELETE ROLE
     @DeleteMapping("/{roleId}")
-    public String deleteRole(@PathVariable Long roleId) {
+    public String deleteRole(@PathVariable Integer roleId) {
         roleService.deleteRole(roleId);
         return "Role deleted successfully";
     }
